\name{baseccirpp09}
\alias{baseccirpp09}
\docType{data}
\title{ of a imagimary class}
\description{

}
\usage{data(baseccirpp09)}
\format{
This table contains about 36,000 observations from French towns. We are interested in the following columns
\describe{
  \item{REG}{Region}
  \item{CODGEO}{Town ID}
  \item{NBFF09}{Number of  households in 2009}
   \item{IMPTOTFF09}{Total taxes collected from households in 2009}
  \item{NBFF08}{Number of  households in 2008}
  \item{RNETFF08}{Income declared by  households in 2008}
  }
}
\source{
http://www.insee.fr/fr/ppp/bases-de-donnees/donnees-detaillees/base-cc-irpp-nouv-serie/base-cc-irpp-nouv-serie-09.zip.}
\examples{
data(baseccirpp09)
}
\keyword{datasets}
